jsp 추가

+++ src/main/java/com/example/demo/controller/UsrArticleController.java
@@ -11,6 +11,8 @@ import com.example.demo.util.Util;
 import com.example.demo.vo.Article;
 import com.example.demo.vo.ResultData;
 
+import jakarta.servlet.http.HttpSession;
+
 @Controller
 public class UsrArticleController {
 	private ArticleService articleService;
@@ -21,7 +23,12 @@ public class UsrArticleController {
 
 	@GetMapping("/usr/article/doWrite")
 	@ResponseBody
-	public ResultData<Article> doWrite(String title, String body) {
+	public ResultData<Article> doWrite(HttpSession session, String title, String body) {
+		
+		if (session.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-L", "로그인 후 이용해주세요");
+		}
+		
 		
 		if (Util.isEmpty(title))
 			return ResultData.from("F-1", "제목을 입력해주세요.");
@@ -29,7 +36,7 @@ public class UsrArticleController {
 		if (Util.isEmpty(body))
 			return ResultData.from("F-2", "내용을 입력해주세요.");
 
-		articleService.writeArticle(title, body);
+		articleService.writeArticle((int) session.getAttribute("loginedMemberId"), title, body);
 		
 		int id = articleService.getLastInsertId();
 		
@@ -52,7 +59,7 @@ public class UsrArticleController {
 	@ResponseBody
 	public ResultData<Article> showDetail(int id) {
 
-		Article foundArticle = articleService.getArticleById(id);
+		Article foundArticle = articleService.forPrintArticle(id);
 
 		if (foundArticle == null) 
 			return ResultData.from("F-1", String.format("%d 번 게시물은 존재하지 않습니다.", id));
@@ -62,12 +69,20 @@ public class UsrArticleController {
 
 	@GetMapping("/usr/article/doModify")
 	@ResponseBody
-	public ResultData doMoidfy(int id, String title, String body) {
+	public ResultData doMoidfy(HttpSession session, int id, String title, String body) {
+		
+		if (session.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-L", "로그인 후 이용해주세요");
+		}
 
 		Article foundArticle = articleService.getArticleById(id);
 
 		if (foundArticle == null) 
 			return ResultData.from("F-1", String.format("%d 번 게시물은 존재하지 않습니다.", id));
+		
+		if ((int) session.getAttribute("loginedMemberId") != foundArticle.getMemberId()) {
+			return ResultData.from("F-A", "게시글 권한이 없습니다.");
+		}
 
 		articleService.modifyAricle(id, title, body);
 		return ResultData.from("S-1", String.format("%d 번 게시물을 수정했습니다.", id));
@@ -75,12 +90,20 @@ public class UsrArticleController {
 
 	@GetMapping("/usr/article/doDelete")
 	@ResponseBody
-	public ResultData doDelete(int id) {
+	public ResultData doDelete(HttpSession session, int id) {
+		
+		if (session.getAttribute("loginedMemberId") == null) {
+			return ResultData.from("F-L", "로그인 후 이용해주세요");
+		}
 
 		Article foundArticle = articleService.getArticleById(id);
 
 		if (foundArticle == null) {
 			return  ResultData.from("F-1", String.format("%d 번 게시물은 존재하지 않습니다.", id));
+		}
+		
+		if ((int) session.getAttribute("loginedMemberId") != foundArticle.getMemberId()) {
+			return ResultData.from("F-A", "게시글 권한이 없습니다.");
 		}
 
 		articleService.deleteArticle(foundArticle);
